requirements.txt
azure-identity==1.14.0
Flask==2.3.2
openai==1.13.3
azure-search-documents==11.4.0b6
azure-storage-blob==12.17.0
python-dotenv==1.0.0
azure-cosmos==4.5.0
opencensus_ext_azure==1.1.11